Refactor the scim2 client
Make as OSGI bundle.
Change the license headers.
