+++ pom.xml
@@ -19,7 +19,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.wso2.identity.client</groupId>
     <artifactId>org.wso2.identity.client.scim2</artifactId>
-    <packaging>jar</packaging>
+    <packaging>bundle</packaging>
     <name>identity-client-scim2</name>
     <version>1.0.0</version>
 
@@ -41,6 +41,35 @@
 
     <build>
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.felix</groupId>
+                <artifactId>maven-bundle-plugin</artifactId>
+                <extensions>true</extensions>
+                <configuration>
+                    <instructions>
+                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
+                        <Bundle-Name>${project.artifactId}</Bundle-Name>
+                        <Export-Package>
+                            org.wso2.scim2.*;version=${project.version},
+                            io.scim2.swagger.client.*;version=${project.version},
+                            com.squareup.okio.*,
+                            com.squareup.okhttp.*,
+                            com.google.gson.*;version=${gson-version},
+                            org.slf4j.*;version=${sl4j-version},
+                            org.joda.time.*;version=${jodatime-version}
+                        </Export-Package>
+                        <Import-Package>
+                            io.swagger.*;version=${swagger-core-version},
+                            org.wso2.charon3.core.*; version="[3.0.0,4.0.0)",
+                            *;resolution:=optional
+                        </Import-Package>
+                    </instructions>
+                </configuration>
+            </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
@@ -72,22 +101,22 @@
                 </executions>
             </plugin>
 
-            <!-- attach test jar -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <version>2.2</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>jar</goal>
-                            <goal>test-jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                </configuration>
-            </plugin>
+            <!--&lt;!&ndash; attach test jar &ndash;&gt;-->
+            <!--<plugin>-->
+                <!--<groupId>org.apache.maven.plugins</groupId>-->
+                <!--<artifactId>maven-jar-plugin</artifactId>-->
+                <!--<version>2.2</version>-->
+                <!--<executions>-->
+                    <!--<execution>-->
+                        <!--<goals>-->
+                            <!--<goal>jar</goal>-->
+                            <!--<goal>test-jar</goal>-->
+                        <!--</goals>-->
+                    <!--</execution>-->
+                <!--</executions>-->
+                <!--<configuration>-->
+                <!--</configuration>-->
+            <!--</plugin>-->
 
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
@@ -177,7 +206,7 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-log4j12</artifactId>
-            <version>1.7.25</version>
+            <version>${sl4j-version}</version>
         </dependency>
         <dependency>
             <groupId>io.swagger</groupId>
@@ -209,6 +238,12 @@
             <artifactId>org.wso2.charon3.core</artifactId>
             <version>${charon.version}</version>
         </dependency>
+        <dependency>
+            <groupId>com.squareup.okio</groupId>
+            <artifactId>okio</artifactId>
+            <version>1.9.0</version>
+            <scope>provided</scope>
+        </dependency>
         <!-- test dependencies -->
         <dependency>
             <groupId>junit</groupId>
@@ -226,6 +261,7 @@
         <okhttp-version>2.7.5</okhttp-version>
         <gson-version>2.8.1</gson-version>
         <jodatime-version>2.9.9</jodatime-version>
+        <sl4j-version>1.7.25</sl4j-version>
         <charon.version>3.0.0</charon.version>
         <commons.logging.version>1.1</commons.logging.version>
         <maven-plugin-version>1.0.0</maven-plugin-version>
