+++ src/main/java/io/scim2/swagger/client/api/Scimv2BulkApi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package io.swagger.client.api;
+package io.scim2.swagger.client.api;
 
 import com.google.gson.reflect.TypeToken;
 import com.squareup.okhttp.Call;
-import io.swagger.client.ApiClient;
-import io.swagger.client.ApiException;
-import io.swagger.client.ApiResponse;
-import io.swagger.client.Configuration;
+import io.scim2.swagger.client.ApiClient;
+import io.scim2.swagger.client.ApiException;
+import io.scim2.swagger.client.ApiResponse;
+import io.scim2.swagger.client.Configuration;
 
 import java.lang.reflect.Type;
 import java.util.List;
