+++ src/main/java/io/scim2/swagger/client/api/Scimv2ServiceProviderConfigApi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,56 +14,54 @@
  * limitations under the License.
  */
 
-package io.swagger.client.api;
+package io.scim2.swagger.client.api;
 
 import com.google.gson.reflect.TypeToken;
 import com.squareup.okhttp.Call;
-import io.swagger.client.ApiClient;
-import io.swagger.client.ApiException;
-import io.swagger.client.ApiResponse;
-import io.swagger.client.Configuration;
+import io.scim2.swagger.client.Configuration;
+import io.scim2.swagger.client.ApiClient;
+import io.scim2.swagger.client.ApiException;
+import io.scim2.swagger.client.ApiResponse;
 
 import java.lang.reflect.Type;
 
-public class Scimv2ResourceTypeApi extends Scimv2BaseApi {
+public class Scimv2ServiceProviderConfigApi extends Scimv2BaseApi {
 
-    public Scimv2ResourceTypeApi() {
+    public Scimv2ServiceProviderConfigApi() {
 
         this(Configuration.getDefaultApiClient());
     }
 
-    public Scimv2ResourceTypeApi(ApiClient apiClient) {
+    public Scimv2ServiceProviderConfigApi(ApiClient apiClient) {
 
         super(apiClient);
     }
     
-    private Call getResourceTypeValidateBeforeCall() throws ApiException {
-        
+    private Call getServiceProviderConfigValidateBeforeCall() throws ApiException {
+
         Call call = getResourceTypeCall();
         return call;
     }
 
     /**
-     * Return the ResourceType schema.
+     * Return the ServiceProviderConfig schema.
      * Returns HTTP 200 if the schema is found.
      * @return String
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public String getResourceType() throws ApiException {
-
-        ApiResponse<String> resp = getResourceTypeWithHttpInfo();
+    public String getServiceProviderConfig() throws ApiException {
+        ApiResponse<String> resp = getServiceProviderConfigWithHttpInfo();
         return resp.getData();
     }
 
     /**
-     * Return the ResourceType schema.
+     * Return the ServiceProviderConfig schema.
      * Returns HTTP 200 if the schema is found.
      * @return ApiResponse&lt;String&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<String> getResourceTypeWithHttpInfo() throws ApiException {
-
-        Call call = getResourceTypeValidateBeforeCall();
+    public ApiResponse<String> getServiceProviderConfigWithHttpInfo() throws ApiException {
+        Call call = getServiceProviderConfigValidateBeforeCall();
         Type localVarReturnType = new TypeToken<String>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
     }
