+++ src/main/java/io/scim2/swagger/client/auth/ApiKeyAuth.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package io.swagger.client.auth;
+package io.scim2.swagger.client.auth;
 
-import io.swagger.client.Pair;
+import io.scim2.swagger.client.Pair;
 
 import java.util.Map;
 import java.util.List;
