+++ src/main/java/io/scim2/swagger/client/auth/HttpBasicAuth.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package io.swagger.client.auth;
+package io.scim2.swagger.client.auth;
 
 import com.squareup.okhttp.Credentials;
-import io.swagger.client.Pair;
+import io.scim2.swagger.client.Pair;
 
 import java.util.List;
 import java.util.Map;
