+++ src/main/java/org/wso2/scim2/operation/GroupOperations.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,9 @@
 
 package org.wso2.scim2.operation;
 
-import io.swagger.client.ApiException;
-import io.swagger.client.ApiResponse;
-import io.swagger.client.api.Scimv2GroupsApi;
+import io.scim2.swagger.client.ApiException;
+import io.scim2.swagger.client.ApiResponse;
+import io.scim2.swagger.client.api.Scimv2GroupsApi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.charon3.core.exceptions.AbstractCharonException;
